Enable compilation of app_kernel
With global CONFIG_NO_OPTIMIZATIONS=y the resulting binary segfaults.

TODO: enable O0 again, but only for the IR step.

"For a complete analysis it is necessary, that all optimizations are switched off.
If this is not done, ARA cannot reliably recognize system calls as
they are potentially affected by inlining."